Don't generate development db

+++ db/seeds.rb
@@ -2,7 +2,7 @@
 Administrator.delete_all
 Administrator.find_or_create_by_name(:name => 'admin', :password => 'foobarbaz', :password_confirmation => 'foobarbaz')
 # For development, use factory girl to build some records.
-#if Rails.env.development?
+if Rails.env.development?
   REDIS.flushall
   Team.delete_all
   Member.delete_all
@@ -25,4 +25,4 @@ Administrator.find_or_create_by_name(:name => 'admin', :password => 'foobarbaz',
     FactoryGirl.create(:design, :team => team)
     team.local_contests << LocalContest.first
   end
-#end
+end
